requirements.txt
# this requirements.txt file is maintained and bumbped by dependabot
# this file is used to see weather new versions of the libraries are available and break shapiq
black==25.1.0
colour==0.1.5
coverage==7.6.12
matplotlib==3.10.1
networkx==3.4.2
pandas==2.2.3
pytest==8.3.4
ruff==0.9.9
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.2
tqdm==4.67.1
torch==2.6.0
torchvision==0.21.0
transformers==4.49.0
tensorflow==2.18.0; python_version < "3.13"
tf-keras==2.18.0; python_version < "3.13"
xgboost==2.1.4
numpy==2.1.2
requests==2.32.3
lightgbm==4.6.0
tabpfn==2.0.6; python_version < "3.13"
